+++ CHANGELOG.md
@@ -10,6 +10,9 @@ omnibus-gitlab repository.
 - Add option to configure postgresql group !2780
 - Add option to configure prometheus group !2782
 - Add option to configure redis group !2784
+11.4.1
+
+- Upgrade Ruby version to 2.4.5
 
 11.4.0